Ensure no breaking changes by keeping legacy Ledger class
